Code Refactor

+++ src/test/java/com/clouway/bank/matchers/DatastoreRule.java
@@ -4,17 +4,13 @@ import com.google.inject.Provider;
 import com.google.inject.util.Providers;
 import com.mongodb.MongoClient;
 import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.MongoIterable;
 import org.junit.rules.ExternalResource;
 
 /**
  * @author Martin Milev <martinmariusmilev@gmail.com>
  */
 public class DatastoreRule extends ExternalResource {
-  private String[] collections;
-
-  public DatastoreRule(String... collections) {
-    this.collections = collections;
-  }
 
   public Provider<MongoDatabase> getDatabase() {
     return Providers.of(new MongoClient("localhost", 27017).getDatabase("testBankApp"));
@@ -22,8 +18,11 @@ public class DatastoreRule extends ExternalResource {
 
   @Override
   protected void before() throws Throwable {
+    MongoIterable<String> collections = getDatabase().get().listCollectionNames();
     for (String collection : collections) {
-      getDatabase().get().getCollection(collection).drop();
+      if(!collection.equals("system.indexes")) {
+        getDatabase().get().getCollection(collection).drop();
+      }
     }
   }
